+++ src/utils/logseq.ts
@@ -9,7 +9,7 @@ export type IBlockNode = {
     children: IBlockNode[],
     properties?: Record<string, any>
     data?: {
-        cursorPosition?: true,
+        setUUID?: string,
         selectionPositions?: number[],
         spawnedBlocks?: IBlockNode[],
         appendedBlocks?: IBlockNode[],