+++ docs/changelog.md
@@ -3,6 +3,9 @@
 - **Query language for pages**:
   - [`.valueType`](reference__query_language.md#filter-value-type) filter with `.onlyStrings` & `.onlyNumbers` shortcuts to filter properties by values' type.
   - [`.journals` & `.day`](reference__query_language.md#filter-journals) filter deal with journal pages.
+- New template tag [`blocks.uuid`](reference__tags_advanced.md#blocks-uuid) for creating block references between template blocks.
+
+  <img width="400px" src="https://github.com/user-attachments/assets/ac0d6a63-f9ff-48f6-9f92-a37d2283a8ca"/>